Refactor FFI interface and document structure
- Removed the FFI module and integrated its functionality into the document and parser modules.
- Updated the main example to demonstrate the new note structure and block handling.
- Introduced a new error handling module to centralize error types across the library.
- Modified the BlockManager to use the new error types for improved error handling.
- Updated models to include DefinitionItem and Attribute for better data representation.
- Changed LeafBlock and ListStyle to use u8 for bullet representation.
- Enhanced the Markdown format handling to support definition lists.
- Updated tests to ensure compatibility with the new error handling and data structures.

+++ src/document/span.rs
@@ -1,12 +1,12 @@
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, uniffi::Record)]
 pub struct TextSpan {
-    pub start: usize,
-    pub end: usize,
+    pub start: u64,
+    pub end: u64,
 }
 
 impl TextSpan {
     #[must_use]
-    pub const fn contains(&self, pos: usize) -> bool {
+    pub const fn contains(&self, pos: u64) -> bool {
         pos >= self.start && pos < self.end
     }
